+++ src/lib/djerba/configure.py
@@ -78,17 +78,18 @@ class configurer(logger):
             self.logger.debug("Found sample names from INI input: {0}".format(samples))
         return samples
 
-    def _compare_coverage_to_target(self,coverage,target):
-        if target > coverage:
-            msg = "Target Depth {0}X is larger than Discovered Coverage {1}X. Changing to Failed mode.".format(target, coverage)
-            self.logger.warning(msg)
-            self.failed = True
-        elif target <= coverage:
-            msg = "Target Depth {0}X is within range of Discovered Coverage {1}X".format(target, coverage)
+    def _compare_coverage_to_target(self):
+        coverage = self.config.getfloat(ini.DISCOVERED, ini.MEAN_COVERAGE)
+        target = self.config.getfloat(ini.DISCOVERED, ini.TARGET_COVERAGE)
+        if coverage >= target:
+            msg = "Mean coverage {0}X is sufficient for target depth of {1}X.".format(coverage, target)
             self.logger.info(msg)
         else:
-            msg = "Target Depth {0}X is incompatible with Discovered Coverage {1}X".format(target, coverage)
-            raise RuntimeError(msg)
+            # if coverage is insufficient, issue a warning
+            # do not force the report to fail; we want the option to proceed if desired
+            msg = "Mean coverage {0}X is NOT sufficient for target depth of {1}X. ".format(coverage, target)+\
+                  "Check coverage and target in config INI are correct. A failed report may be required."
+            self.logger.warning(msg)
 
     def find_data_files(self):
         data_files = {}
@@ -127,17 +128,33 @@ class configurer(logger):
                 updates[ini.GEP_FILE] = self.reader.parse_gep_path()
         updates.update(self.reader.get_sample_names())
         updates.update(self.find_data_files())
+        if self.config.has_option(ini.DISCOVERED, ini.CBIO_STUDY_ID):
+            msg = "Using manually configured cBioportal project name {0}".format(self.config[ini.DISCOVERED][ini.CBIO_STUDY_ID])
+            self.logger.info(msg)
+        else:
+            project_id = self.config[ini.INPUTS][ini.PROJECT_ID]
+            try:
+                cbio_study_id = pull_qc(self.config).fetch_cbio_name(project_id)
+                msg = "Found cbioportal-study-id {0} for project {1} in Shesmu".format(updates[ini.CBIO_STUDY_ID], project_id)
+                self.logger.info(msg)
+            except KeyError:
+                cbio_study_id = project_id
+                msg = "Couldn't find the project in Shesmu. cbioportal-study-id will default to project-id {0}".format(project_id)
+                self.logger.info(msg)
+            updates[ini.CBIO_STUDY_ID] = cbio_study_id
         return updates
 
     def discover_secondary(self):
         updates = {}
+        # configure Sequenza: gamma, solution, purity, ploidy.
+        # Requires Sequenza file from discover_primary()
         reader = sequenza_reader(self.config[ini.DISCOVERED][ini.SEQUENZA_FILE])
         gamma = self.config.getint(ini.DISCOVERED, ini.SEQUENZA_GAMMA, fallback=None)
         solution = self.config.get(ini.DISCOVERED, ini.SEQUENZA_SOLUTION, fallback=None)
         # get_default_gamma_id() returns (gamma, solution)
         if gamma == None:
-                gamma = reader.get_default_gamma_id()[0]
-                self.logger.info("Automatically generated Sequenza gamma: {0}".format(gamma))
+            gamma = reader.get_default_gamma_id()[0]
+            self.logger.info("Automatically generated Sequenza gamma: {0}".format(gamma))
         else:
             self.logger.info("User-supplied Sequenza gamma: {0}".format(gamma))
         if solution == None:
@@ -155,31 +172,47 @@ class configurer(logger):
         updates[ini.SEQUENZA_SOLUTION] = solution
         updates[ini.PLOIDY] = ploidy
         updates[ini.PURITY] = purity
-        tumour_id =  self.config[ini.DISCOVERED][ini.TUMOUR_ID]
-        try:
-            coverage = pull_qc(self.config).fetch_coverage_etl_data(tumour_id)
-        except MissingQCETLError as e:
-            msg = "Can't retrieve coverage associated with tumour_id {0} in QC-ETL. Will use manually configured INI parameter, if available: {1}".format(tumour_id, e)            
-            self.logger.warning(msg)
-            coverage =  self.config[ini.DISCOVERED][ini.MEAN_COVERAGE]
-        try:
-            callability = pull_qc(self.config).fetch_callability_etl_data(tumour_id)
-        except MissingQCETLError as e:
-            msg = "Can't retrieve callability associated with tumour_id {0} in QC-ETL. Will use manually configured INI parameter, if available: {1}".format(tumour_id, e)
-            self.logger.warning(msg)
-            callability = self.config[ini.DISCOVERED][ini.PCT_V7_ABOVE_80X]
-        self.logger.info("Callability {0} and Coverage {1}".format(callability, coverage))
-        updates[ini.PCT_V7_ABOVE_80X] = callability
-        updates[ini.MEAN_COVERAGE] = coverage
-        try:
-            target_coverage = pull_qc(self.config).fetch_pinery_assay(self.config[ini.INPUTS][ini.REQ_ID])
-        except MissingPineryError or UnsupportedAssayError as e:
-            msg = "Can't retrieve target coverage associated with requisition {0} in Pinery. Will use manually configured INI parameter, if available: {1}".format(self.config[ini.INPUTS][ini.REQ_ID], e)
-            self.logger.warning(msg)
-            target_coverage = self.config[ini.DISCOVERED][ini.TARGET_COVERAGE]
-        self.logger.info("Target Coverage: {0}".format(target_coverage))
-        updates[ini.TARGET_COVERAGE] = target_coverage    
-        self._compare_coverage_to_target(coverage,target_coverage)
+        # Configure with QC-ETL and Pinery: mean coverage, target coverage, callability.
+        # Requires tumour and requisition IDs from discover_primary()
+        tumour_id = self.config[ini.DISCOVERED][ini.TUMOUR_ID]
+        qc_retriever = pull_qc(self.config)
+        if self.config.has_option(ini.DISCOVERED, ini.MEAN_COVERAGE):
+            msg = "Using manually configured mean coverage: {0}".format(self.config[ini.DISCOVERED][ini.MEAN_COVERAGE])
+            self.logger.info(msg)
+        else:
+            try:
+                coverage = qc_retriever.fetch_coverage_etl_data(tumour_id)
+            except MissingQCETLError as e:
+                msg = "Coverage not supplied by user, and cannot be retrieved from QC-ETL for tumour_id {0}: {1}".format(tumour_id, e)
+                self.logger.error(msg)
+                raise
+            self.logger.info("Using mean coverage from QC-ETL: {0}".format(coverage))
+            updates[ini.MEAN_COVERAGE] = coverage
+        if self.config.has_option(ini.DISCOVERED, ini.PCT_V7_ABOVE_80X):
+            msg = "Using manually configured callability: {0}".format(self.config[ini.DISCOVERED][ini.PCT_V7_ABOVE_80X])
+            self.logger.info(msg)
+        else:
+            try:
+                callability = qc_retriever.fetch_callability_etl_data(tumour_id)
+            except MissingQCETLError as e:
+                msg = "Callability not supplied by user, and cannot be retrieved from QC-ETL for tumour_id {0}: {1}".format(tumour_id, e)
+                self.logger.error(msg)
+                raise
+            self.logger.info("Using callability from QC-ETL: {0}".format(coverage))
+            updates[ini.PCT_V7_ABOVE_80X] = callability
+        if self.config.has_option(ini.DISCOVERED, ini.TARGET_COVERAGE):
+            msg = "Using manually configured target coverage: {0}".format(self.config[ini.DISCOVERED][ini.TARGET_COVERAGE])
+            self.logger.info(msg)
+        else:
+            req_id = self.config[ini.INPUTS][ini.REQ_ID]
+            try:
+                target_coverage = qc_retriever.fetch_pinery_assay(req_id)
+            except (MissingPineryError, UnsupportedAssayError) as e:
+                msg = "Target coverage not supplied by user, and cannot be retrieved from Pinery for requisition ID {0}: {1}".format(req_id, e)
+                self.logger.error(msg)
+                raise
+            self.logger.info("Using target coverage from Pinery: {0}".format(target_coverage))
+            updates[ini.TARGET_COVERAGE] = target_coverage
         return updates
 
     def run(self, out_path):
@@ -188,10 +221,10 @@ class configurer(logger):
         # first pass -- update basic parameters
         self.update_primary()
         if self.failed:
-            self.logger.info("Failed report mode; omitting sequenza/logR config updates")
-            if not (self.config.has_option(ini.DISCOVERED, ini.PURITY) \
-                    and self.config.has_option(ini.DISCOVERED, ini.PLOIDY)):
-                msg = "Purity/ploidy not found; must be entered manually for failed reports"
+            self.logger.info("Failed report mode; omitting secondary/tertiary config updates")
+            required = [ini.PURITY, ini.PLOIDY]
+            if not all([self.config.has_option(ini.DISCOVERED, x) for x in required]):
+                msg = "The following parameters must be entered manually for failed reports: {0}".format(required)
                 self.logger.error(msg)
                 raise RuntimeError(msg)
         else:
@@ -200,6 +233,7 @@ class configurer(logger):
             self.update_secondary()
             # third pass -- logR cutoffs using the updated purity
             self.update_tertiary()
+            self._compare_coverage_to_target()
         with open(out_path, 'w') as out_file:
             self.config.write(out_file)
         self.logger.info("Djerba config finished; wrote output to {0}".format(out_path))
@@ -296,6 +330,8 @@ class pull_qc(logger):
         self.logger = self.get_logger(log_level, __name__, log_path)
         self.pinery_url = self.config[ini.SETTINGS][ini.PINERY_URL]
         self.qcetl_cache = self.config[ini.SETTINGS][ini.QCETL_CACHE]
+        self.cbio_path = self.config[ini.SETTINGS][ini.CBIO_PROJECT_PATH]
+        path_validator(log_level, log_path).validate_input_file(self.cbio_path)
         self.etl_cache = QCETLCache(self.qcetl_cache)
 
     def fetch_callability_etl_data(self,tumour_id):
@@ -307,11 +343,17 @@ class pull_qc(logger):
             ]
         if len(data) > 0:
             callability = round(data.iloc[0][columns_of_interest.Callability].item() * 100,1)
-            return(callability)
+            return callability
         else:
             msg = "Djerba couldn't find the callability associated with tumour_id {0} in QC-ETL. ".format(tumour_id)
             self.logger.debug(msg)
             raise MissingQCETLError(msg)
+        
+    def fetch_cbio_name(self,project_id):
+        with open(self.cbio_path) as in_file:
+            data = json.loads(in_file.read())
+        cbioportal_project_id = data['values'][project_id]['cbioportal_project']
+        return cbioportal_project_id
         
     def fetch_coverage_etl_data(self,tumour_id):
         cached_coverages = self.etl_cache.bamqc4merged.bamqc4merged
@@ -329,7 +371,12 @@ class pull_qc(logger):
             raise MissingQCETLError(msg)
 
     def fetch_pinery_assay(self,requisition_name: str):
-        pinery_requisition = self.pinery_get(f'/requisition?name={requisition_name}')
+        try:
+            pinery_requisition = self.pinery_get(f'/requisition?name={requisition_name}')
+        except HTTPError:
+            msg = "Djerba couldn't find the assay associated with requisition {0} in Pinery. ".format(requisition_name)
+            self.logger.debug(msg)
+            raise MissingPineryError(msg)
         if len(pinery_requisition) > 0:
             try:
                 pinery_assay = self.pinery_get(f'/assay/{pinery_requisition["assay_id"]}')