+++ src/lib/djerba/plugins/base.py
@@ -10,7 +10,9 @@ from djerba.util.logger import logger
 
 class plugin_base(logger, ABC):
 
-    def __init__(self, log_level=logging.INFO, log_path=None):
+    def __init__(self, workspace, log_level=logging.INFO, log_path=None):
+        # workspace is an instance of djerba.core.workspace
+        self.workspace = workspace
         self.log_level = log_level
         self.log_path = log_path
         self.logger = self.get_logger(log_level, __name__, log_path)