+++ src/lib/djerba/plugins/plugin_tester.py
@@ -13,6 +13,7 @@ import djerba.util.ini_fields as ini
 from configparser import ConfigParser
 from djerba.core.json_validator import plugin_json_validator
 from djerba.core.main import main as core_main
+from djerba.core.workspace import workspace
 from djerba.mergers.gene_information.merger import main as gene_information_merger_main
 from djerba.util.subprocess_runner import subprocess_runner
 from djerba.util.testing.tools import TestBase
@@ -60,7 +61,7 @@ class PluginTester(TestBase):
         expected_md5 = params[self.MD5]
         plugin_name = self.read_plugin_name(ini_path)
         self.assertTrue(plugin_name)
-        djerba_main = core_main(log_level=logging.WARNING)
+        djerba_main = core_main(self.get_tmp_dir(), log_level=logging.WARNING)
         config = djerba_main.configure(ini_path)
         data_found = self.redact_json_data(djerba_main.extract(config))
         with open(expected_json_path) as json_file: