+++ src/lib/djerba/util/testing/tools.py
@@ -4,6 +4,7 @@ Test base class; called 'trial.py' to hide from automated unittest discovery
 
 import hashlib
 import re
+import tempfile
 import time
 import unittest
 import djerba.util.constants as constants
@@ -15,6 +16,11 @@ class TestBase(unittest.TestCase):
     def setUp(self):
         self.path_validator = path_validator()
         self.maxDiff = None
+        self.tmp = tempfile.TemporaryDirectory(prefix='djerba_')
+        self.tmp_dir = self.tmp.name
+
+    def tearDown(self):
+        self.tmp.cleanup()
 
     def assert_report_MD5(self, report_string, expected_md5):
         body = self.redact_html(report_string)
@@ -29,6 +35,10 @@ class TestBase(unittest.TestCase):
         md5 = hashlib.md5()
         md5.update(input_string.encode(constants.TEXT_ENCODING))
         return md5.hexdigest()
+
+    def get_tmp_dir(self):
+        # convenience method; returns path to the temp dir
+        return self.tmp_dir
 
     def redact_html(self, report_string):
         """